Allow setting of SOA bot's aggressiveness
Currently the SOA bot doesn't have a way to control aggressiveness in
the stance extraction module. We add this capability by implementing
an enum called aggressiveness with three modes "aggressive," "moderate,"
and "friendly." It affects aggressiveness by adjusting the hyperparameters
of the greek letters in the stance extraction module function.