Add script to download game logs
`run_bots.py` currently outputs a JSON representation of the game state
to a file. However, this file cannot be accessed when the script is run
in a container. In addition, the output only contains the game state as
viewable by the agent itself.

The complete game state is available through the `diplomacy` interface,
so it is much simpler to retrieve it directly from the source.