Refresh local copy of orders to include replacements
This should fix an observed bug where the SOA bot would attack allies.

The SOA bot stores planned orders both in a variable local to the
`__call__()` method and in an instance variable for the entire object.
There is some copying back and forth between the two, but the existing
code leaves the local variable out of date by the time it's used as a
return value for the method. Although a larger refactoring should be
done to prevent this problem, this commit adds a statement to update the
local variable after the instance variable is updated through another
method.