Prevent sending duplicate messages
The messaging logic was written assuming it would be called only once,
so the addition of the messaging loop caused messages to be sent up to 3
times per phase.

`MessagesData.add_message()` now has an option (disabled by default for
backwards compatibility) to only store new messages and to return
whether the message was added or not. This return value is used in
`SmartOrderAccepterBot.send_message()` to only send messages that are
new.

As part of this change, `MessagesData` now implements
`collections.abc.Collection` to have a more useable interface. In
addition, it now has a custom `__str__()` method for more useful debug
output.