Disable `ORR` in SOA bot with `DISABLE_ORR` env var
This is a hacky way to work around Cynnovative not supporting the `ORR`
predicate from DAIDE. It overrides two imported functions to return
without adding an `ORR`. This code can hopefully be removed in the
future.

Setting `DISABLE_ORR` to any value (even the empty string) will disable
`ORR` generation.