Add much more intent logging
Some of these were converted from print statements we already had.
Others are new.

As part of this, I carried out a slight refactor to the code for
generating an intent message for power relations.

The stance vector we print to the console now uses fewer significant
digits because full precision was not useful.