Refactor Dockerfile to reduce container size
I recently found Hadolint, a tool for linting Dockerfiles, and decided
to run it on our existing file. It found many possible ways for it to be
improved, which mostly consisted of combining `RUN` statements to save
space and pinning versions for reproducibility. I also removed some
actions that were unneeded and deleted some files that do not need to be
in the final image.

The results are impressive. When building `allan_dip_bot` from
`pcpaquette/tensorflow-serving`, the container size went from 4.12 GB to
3.47 GB. In addition, the build time went from 2m33s to 2m7s.

I also modified the build instructions so we can reuse build caches
across machines. This will further increase reproducibility and reduce
build and transfer times for containers.