Define `PYTHONPATH` earlier in `Dockerfile`
`PYTHONPATH` was only being declared in the `dev` target, so the
production `allan_dip_bot` did not have it defined. This was not a
problem because it is defined in `run.sh`, but adding `set -u` in a
later commit caused the script to fail. Defining `PYTHONPATH` in the
`base` target prevents that failure.