Set `safe.directory` at system (not user) level
This is needed for Singularity to work properly because Docker
constructs the container as `root`, but Singularity runs the container
as the user running it. The `.gitignore` file is owned by `root` and
therefore cannot be used by the unprivileged user. The solution is to
set `/etc/gitconfig` to solve the problem for all users within the
container.