Do not send VIA orders from random proposer bot
We need to convert DipNet orders to DAIDE before sending them, and DAIDE
requires that a movement via a convoy lists the entire path. When orders
are randomly generated, the chance of an appropriate convoy existing is
miniscule, so we cannot do the conversion properly. Avoiding proposing
VIA orders is the simplest solution to this problem.