Generate game IDs in `test_send_message()` with fewer collisions
We were generating a game ID with a random suffix as part of the
`test_send_message()` test to ensure `NetworkGame`s work properly. While
looking at the console output from an unrelated failure, I noticed that
we had 156 collisions before we found an unused ID. I initially
increased the maximum random suffixed number from 10,000 to 1 billion,
but I realized that appending the current date with microsecond
precision should be sufficient to avoid collisions. It also has the
advantage of making it easier to browse created games if needed.