Consolidate message retrieval code
We had various forms of the message retrieval code scattered throughout
the codebase, so it made sense to merge them together and store the
resulting function in a single place.

A major change here is that the bots can now read messages on-demand
instead of having them passed in. The SOA bot was already ignoring the
messages passed in, but this removes the need for specialized SOA bot
code.

This change revealed a bug where we would generate invalid DAIDE during
some message processing. This has been fixed by filtering out any
messages sent by the bots so they do not try to respond to themselves.