Fix messaging in pushover bot
There was an `if` statment with the condition `"FCT" not in
last_message.message`. However, a previous `if` statement in the method
would return when `"FCT" in last_message.message`, so the body of the
later `if` statement would never run. Given that the `if` statement was
in a loop, I changed `last_message` to the loop variable `message`,
which is what I presume the original intention was.