Simplify `BaselineBots` interface
We do not need `gen_messages()` and `gen_orders()` methods for every
bot, so there is no reason to enforce it. The only method being called
externally is `__call__()`, so it should be abstract and the other
methods should be removed.

I have removed or changed the interfaces of the methods that are no
longer required to simplify their respective bots.