Consolidate message sending code
We had various forms of the message sending code scattered throughout
the codebase, so it made sense to merge them together and store the
resulting function in a single place.

A major change here is that the bots can now send messages on-demand
instead of needing to return them and have them all sent at once when
the round ended. The SOA bot was already sending messages on-demand, but
this removes the duplicate sending of its messages in our test fixture.