Use relative path for download script in `run_games.sh`
Displaying the full path in the command to run is excessive. The
relative path is only used if the download script is contained (either
directly or as a descendent) of the current working directory.

As part of this, I factored out some of the code for path handling to
reduce repetition.