Fix test failures in `parse_proposal_messages()`
These failures only occur when the exceptions that are usually caught in
the function are not.

I added another check to `test_parse_proposals()` to make sure that the
number of input and output messsages to `parse_proposal_messages()` is
the same. While not true in general, it is expected in this test case,
and the test should fail if the function unexpectedly drops messages.