Add time-bounded messaging loop for SOA bot
When playing against other bots, we don't know anything about their
communication timing. To compensate, I have added the option to have our
messaging continue until we run out of time. This new behavior is the
default, and the old behavior of running a set number of iterations has
been kept for easier unit testing. The time allotted for the messaging
loop is controllable with the new `--communication_stage_length`
argument.

We reraise `CancelledError` from all `try` blocks because
`CancelledError` unintuitively inherits from `Exception` instead of
`BaseException`. This makes it easy to accidentally catch and ignore,
preventing the task from being cancelled. Python 3.8 wisely changes
this, but that doesn't matter for us.