Raise extra exceptions in new debug mode
These are normally ignored for robustness, but they should be raised
during development so they are not missed. These exceptions can be
enabled by setting the `ALLAN_DEBUG` environment variable.

There is one ignored exception that isn't being reraised here, but that
`try` statement is not for unexpected behavior like these are.