Simplify image building
Now that building is much simpler and the resulting images are *much*
smaller, we no longer need to rely on the remote build cache as heavily.
This allows us to simplify image building.

I also renamed the build target to `achilles`, which is what I have been
calling this new framework in my head. In Greek mythology, Achilles was
the pupil of the centaur Chiron.