Remove `docs/` and related code
The documentation is severely outdated, and no one is using it. Instead
of having a large amount of misleading information, I am just removing
it.

I have also removed some related code, such as the dependencies needed
to build the docs and the configuration needed to handle the `.bat` file
properly.