Create `RandomProposerAdvisor` bot
Instead of a `RandomProposerBot`, there is now a `RandomProposerAdvisor`
and a `RandomProposerPlayer`. They share the same core logic, but the
separate classes make it possible to select which type is wanted as
runtime without passing any new parameters.

This is a pretty substantial change that took a lot of experimentation
to develop properly, and I could think of no better way to arrange the
commits than to simply squash them all together.